chore: componentize availability component
